Add submit description file for Aggregating Evidence
EPoD's blended learning trainings almost always take place internationally.
This means that if learner data is to be made available for instructor
review as soon as possible after completion of EPoDx digital units, the
process of downloading and uploading learner data must be automated. To do
so, we take advantage of the Harvard-MIT Data Center (HMDC) Research
Computing Environment (RCE). The RCE has servers dedicated to batch
processing that run on a high throughput cluster, based on HTCondor, which
allows jobs to be "deferred" so that they are run at scheduled future
dates and times or at specified intervals. This submit description file
defines parameters so that write_to_gsheets_AGG.py is run at a scheduled
date and time or at a specified interval.