Create single script and submit file for all units.
Previously, I had written separate Python scripts and submit files for
each EPoDx unit. This commit updates the write_to_g_sheet function so
that the same script can be used to write to the master sheets for all
six EPoDx units. This commit also renames the script and submit files
to reflect this change.