Add function to tunnel and write to sheet
Previously, I had chosen a large value (5 min) for the sleep command
when establishing the SSH tunnel but this does not guarantee that the
tunnel will stay open long enough to write to all six sheets. Instead,
this commit defines a new function which establishes a new tunnel for
writing to each sheet.