Enable looping over multiple dashboards per course
This is a more permanent fix than the one implemented yesterday. The
single argument for tunnel_and_write_to_g_sheet is now a list where the
first item is the course ID and the second item is a sheet identifier.
This allows for looping over a list of lists.