Add argument to allow writing only profile or problem data.
When instructors want access to dashboards less than 24 hours after a
cohort was enrolled, fully automatic updating is not possible because
the learner profiles data on the edX Analytics API only updates every 24
hours. However, learner profile data can be generated on the fly from
the instructor dashboard of epodx.org and manually uploaded. In cases
where the learner profile data for a cohort is static, such as when the
cohort was enrolled in a computer lab, it is helpful to set up the
problem response data to automatically update without overwriting the
learner profiles data that had been manually uploaded. To accomplish
this, an argument is added to the write_to_g_sheet function which allows
specification of whether to download and write both learner profiles and
problem responses data, just learner profiles, or just problem
responses.