Automate learner engagement data downloads
EPoDx tracks weekly learner engagement including problem attempts,
problem submissions, and video views. Because it might be useful for
future research, we want to archive this data. I had previously been
doing this through the insights.epodx.org GUI. This, however, had two
problems. 1) It was manual. 2) I didn't realize that the csv downloaded
via the "Download CSV" button did not match the GUI display in that it
included the usernames of active learners but not their engagement
statistics.

This script fixes these two problems by automating the process of
downloading engagement data via the API and writing it to an archive
folder in Dropbox.

In future, I would like to modify the script so that it appends new data
each week rather than creating a separate file.