Specify ignore_segments parameter
The first version of this script specified to pull learner engagement
data for the following segments, "highly_engaged," "disengaging," and
"struggling." However, having discovered that the segment parameter was
blank for one learner who engaged with the unit this past week, I have
changed the specification to "ignore_segments: 'inactive'" (which has
the added advantage of being more concise). I'll follow up with Jill to
ask why there are missing values.