Add NSPP dashboard for SYS
Trainings for the SMC 23 cohort are covering both Systematic Approaches
and Impact Evaluations. To give local instructors dedicated dashboards,
I have created a separate set of Google Sheets for the Systematic
Approaches unit. This is in addition to the set of Google Sheets for
NSPP that already existed for Impact Evaluations.