Add note that 32 GB memory requirement is indicative of bug
The script that the job runs connects to the edX analytics API,
downloads two files (average size ~350 KB) and then writes each to a
separate Google Sheet using the Sheets API. Nonetheless, I had been
getting an error on the RCE that the job was put on hold for exceeding
its memory limit of 32 GB.

I have not debugged this error since, for other reasons, I shifted to
asking local partners to manually do the updating. When automation was
desired, I used my local crontab as a workaround.