Refactor: Update write_to_gsheet function arguments
In the previous version, the write_to_gsheet function asked users to
specify whether writing to the "primary" or "secondary" dashboard and
selected the spreadsheet ID through a long series of if-else statements.

This refactoring updates the function to reflect the fact that sheets
are now organized according to partners and defines spreadsheet ID by
referencing a dictionary in secrets.py.